refactor: split gateway core read models
